+++ src/golang/time/index.ts
@@ -1,0 +1,1 @@
+export { Format } from './Format.ts'